CategoryAdapter changes - Now categories reacts on clicks - Deleted RecyclerViewClickInterface.java - Changed category_card.xml layout - Small changes in PasswordsFragment.java - Deleted useless import in RegistrationException.java - Moved all adapters to package "adapters"
