+++ app/src/main/java/com/lovejazz/kyle/PasswordsFragment.java
@@ -1,6 +1,5 @@
 package com.lovejazz.kyle;
 
-import android.content.Intent;
 import android.os.Bundle;
 import android.util.Log;
 import android.view.LayoutInflater;
@@ -20,6 +19,7 @@ import com.google.firebase.auth.FirebaseAuth;
 import com.google.firebase.firestore.FirebaseFirestore;
 import com.google.firebase.firestore.QueryDocumentSnapshot;
 import com.google.firebase.firestore.QuerySnapshot;
+import com.lovejazz.kyle.adapters.CategoryAdapter;
 
 import java.util.AbstractMap;
 import java.util.ArrayList;
@@ -31,7 +31,7 @@ import java.util.TreeMap;
 
 import static com.lovejazz.kyle.EntryUtils.makeSnackbarError;
 
-public class PasswordsFragment extends Fragment implements RecyclerViewClickInterface {
+public class PasswordsFragment extends Fragment {
     private static final String TAG = "c";
     private FirebaseFirestore fstore;
     private FirebaseAuth mAuth;
@@ -249,19 +249,9 @@ public class PasswordsFragment extends Fragment implements RecyclerViewClickInte
         }
         RecyclerView recyclerView = view.findViewById(R.id.categories_recycler);
         CategoryAdapter categoryAdapter = new CategoryAdapter(getContext(), categoryNames,
-                linksToIcons, linkToBackgroundImages, this);
+                linksToIcons, linkToBackgroundImages);
         recyclerView.setAdapter(categoryAdapter);
-        GridLayoutManager gridLayoutManager = new GridLayoutManager(getContext(),
-                2, GridLayoutManager.VERTICAL, false);
+        GridLayoutManager gridLayoutManager = new GridLayoutManager(getContext(), 2);
         recyclerView.setLayoutManager(gridLayoutManager);
-    }
-
-    @Override
-    public void onItemClick(int position) {
-        String categoryName = categoriesList.get(position).getName();
-        Log.d(TAG, position + " - position");
-        Intent intentToCategoryActivity = new Intent(getContext(), CategoryActivity.class);
-        intentToCategoryActivity.putExtra("categoryName", categoryName);
-        startActivity(intentToCategoryActivity);
     }
 }
