--- app/src/main/java/com/lovejazz/kyle/RecyclerViewClickInterface.java
@@ -1,5 +1,0 @@
-package com.lovejazz.kyle;
-
-public interface RecyclerViewClickInterface {
-    void onItemClick(int position);
-}
