+++ app/src/main/java/com/lovejazz/kyle/adapters/AccountAdapter.java
@@ -1,18 +1,18 @@
-package com.lovejazz.kyle;
+package com.lovejazz.kyle.adapters;
 
-import android.content.Context;
 import android.graphics.drawable.Drawable;
 import android.view.LayoutInflater;
-import android.view.View;
 import android.view.ViewGroup;
 import android.widget.ImageView;
 import android.widget.TextView;
 
 import androidx.annotation.NonNull;
 import androidx.cardview.widget.CardView;
-import androidx.constraintlayout.widget.ConstraintLayout;
 import androidx.core.content.ContextCompat;
 import androidx.recyclerview.widget.RecyclerView;
+
+import com.lovejazz.kyle.Account;
+import com.lovejazz.kyle.R;
 
 public class AccountAdapter extends RecyclerView.Adapter<AccountAdapter.ViewHolder> {
     private String[] names;
